INode.MapTerms extension
